added tagClass to options
with this new option we can add a class to the tags themselves when they get added to the container or before the input element. This would make it so we dont need to include the default css that came with tag manager, along with allowing us to use predefined css rules to the tags.
In my case, I can now do the following with this addition..
````javascript
$('.tags').tagsManager({
  tagClass: 'label label-inverse'
})
````
and the newly added tags will have the label classes applied to it.