Flexible "Base Delimiter" (Issue #104)
Previously TagManager assumes it uses a comma as the base delimiter for the hidden field. This commit enables the use of other delimiters, by using the first char (non-key) value of the existing 'delimiters' parameter.

If no char delimiter is specified in the 'delimiters' parameter, comma will be used by default.

Includes documentation update.